requirements.txt
# Core Dependencies
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# Audio Processing
pyaudio>=0.2.11
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.1

# Computer Vision & OCR
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Windows System Integration
pywin32>=306
psutil>=5.9.0
pyautogui>=0.9.54
pynput>=1.7.6
win32gui>=221
win32api>=221
win32con>=221

# Web Automation
selenium>=4.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Natural Language Processing
spacy>=3.6.0
nltk>=3.8.1
textblob>=0.17.1

# GUI Framework
tkinter
customtkinter>=5.2.0
PyQt6>=6.5.0

# Financial Data
yfinance>=0.2.18
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0

# Configuration & Logging
pyyaml>=6.0
configparser>=5.3.0
logging>=0.4.9.6

# API Integration
openai>=0.27.0
anthropic>=0.3.0
httpx>=0.24.0

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0